% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_timeseries.R
\name{getCalendarDayMeans}
\alias{getCalendarDayMeans}
\title{Calculate calendar day mean temperatures from historical series}
\usage{
getCalendarDayMeans(
  temp_hist = temp_hist,
  lookback = "-10 year",
  series_ending = NULL
)
}
\arguments{
\item{temp_hist}{Temperature time series formatted using
\code{\link[mosqmod]{formatTempSeq}}.}

\item{lookback}{Subset the historical time series starting at a past date.
Must match the format for the 'by' argument in \code{\link{seq.Date}}.}

\item{series_ending}{Subset the historical time series ending date. Must be a
valid \code{\link{Dates}} object. When set to NULL (default) the most
recent date in the temperature time series is used.}
}
\description{
Calculate calendar day mean temperatures from historical series
}
\examples{
## get past temperature series from append_TempSeq example
example("append_TempSeq")
temp_hist <- mosqmod::formatTempSeq(temp_hist)
# using default - 10 years from most recent temperature
getCalendarDayMeans(temp_hist = temp_hist)
}
