% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosqpopn.R
\name{mosqpopn}
\alias{mosqpopn}
\title{Title}
\usage{
mosqpopn(
  temp_ts,
  b,
  alpha,
  beta,
  K_L,
  M_max,
  MTD,
  L_1 = 0,
  L_2 = 0,
  L_3 = 0,
  L_4 = 0,
  L_5 = 0,
  M = 0,
  Mfloor = 100
)
}
\arguments{
\item{temp_ts}{temperature time series}

\item{b}{number of female eggs per clutch}

\item{alpha}{adult mortality rate (1/days)}

\item{beta}{larval mortality rate (1/days)}

\item{K_L}{larval carrying capacity (numbers/km^2)}

\item{M_max}{max adult density}

\item{MTD}{minimum temperature for mosquito development)}

\item{L_1}{Initial instar densities (1-5)}

\item{L_2}{see above}

\item{L_3}{see above}

\item{L_4}{see above}

\item{L_5}{see above}

\item{M}{Initial adult density}
}
\value{

}
\description{
Title
}
