% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_popn}
\alias{plot_popn}
\title{Population size time series}
\usage{
plot_popn(
  resdf,
  selectPopn = c("L", "L_1", "L_2", "L_3", "L_4", "L_5", "M"),
  include_temp = TRUE,
  include_plasmod = TRUE,
  MTT = NULL,
  MTD = NULL
)
}
\arguments{
\item{resdf}{Results dataframe generated by mosqmod::runModel().}

\item{selectPopn}{Character vector of population component to plot (options are "L", "L_1", "L_2", "L_3", "L_4", "L_5" or "M").}

\item{include_temp}{Include a plot facet of temperature time series (default = TRUE).}

\item{include_plasmod}{Include (default = TRUE) shaded regions on plot when
temperatures are suitable for plasmodium development.}

\item{MTT}{Display minimum temperature threshold for mosquito development.}

\item{MTD}{Display minimum temperature threshold for plasmodium sporogeny.}
}
\description{
Plot time series of population size using model outputs.
}
