% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_popn_years}
\alias{plot_popn_years}
\title{Population size time series overlay}
\usage{
plot_popn_years(
  resdf,
  selectPopn = c("L", "L_1", "L_2", "L_3", "L_4", "L_5", "M")
)
}
\arguments{
\item{resdf}{Results dataframe generated by mosqmod::runModel().}

\item{selectPopn}{Character vector of population component to plot (options are "L", "L_1", "L_2", "L_3", "L_4", "L_5" or "M").}
}
\value{

}
\description{
Population size time series of model outputs with years overlaid.
}
