% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosqpopn.R
\name{runMosqModApp}
\alias{runMosqModApp}
\title{Run shiny app locally}
\usage{
runMosqModApp(cliflo_requests = TRUE)
}
\arguments{
\item{cliflo_requests}{Make requests to CliFlo database to update
temperatures (default = TRUE). Set to FALSE to disable when working on
restricted networks for example.}
}
\value{

}
\description{
Runs shiny app locally using the 'app/app.R' file included in 'mosqmod'
package (i.e. 'Program Files/R/R.X.X.X/library/mosqmod').
}
\examples{
mosqmod::runMosqModApp(cliflo_requests = FALSE)
}
