% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saved_station_temps}
\alias{saved_station_temps}
\title{Stored temperature time series}
\format{
An object of class \code{cfTemp} with 15899 rows and 12 columns.
}
\usage{
saved_station_temps
}
\description{
A dataset of daily stored temperatures from Musselburgh and Nugget Point
climate stations and accessed using the clifro package in R. See the
'update_stored_temps.R' script in the data-raw folder of this package for
details on updating temperature values.
}
\keyword{datasets}
