% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_timeseries.R
\name{append_TempSeq}
\alias{append_TempSeq}
\title{Append temperature times series with latest temperature from NIWA's CliFlo database.}
\usage{
append_TempSeq(
  temp_stored,
  date_append = as.Date(Sys.time()),
  request = TRUE,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{temp_stored}{Stored temperatures to append.}

\item{date_append}{Get updated temperatures after this date. Defaults to current system clock.}

\item{request}{Make CliFlo request (default = TRUE).}

\item{username}{CliFlo login username (see 'https://cliflo.niwa.co.nz')}

\item{password}{CliFlo login password.}
}
\description{
Append temperature times series with latest temperature from NIWA's CliFlo database.
}
\examples{
# get stored data
dat <- data.frame(mosqmod::saved_station_temps)
# subset selected temperature series
subdat <- subset(dat, Station == "Dunedin, Musselburgh Ews")

temp_hist <- mosqmod::append_TempSeq(temp_stored = subdat,
                                     request = FALSE,
                                     date_append = as.Date(Sys.Date()))
}
