% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_popn}
\alias{plot_popn}
\title{Population size time series}
\usage{
plot_popn(
  resdf,
  selectPopn = c("L", "L_1", "L_2", "L_3", "L_4", "L_5", "M"),
  include_temp = TRUE
)
}
\arguments{
\item{resdf}{Results dataframe generated by mosqmod::runModel().}

\item{selectPopn}{Character vector of population component to plot (options are "L", "L_1", "L_2", "L_3", "L_4", "L_5" or "M").}

\item{include_temp}{Include a plot facet of temperature time series (default = TRUE).}
}
\value{

}
\description{
Plot time series of population size using model outputs.
}
