% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosqpopn.R
\name{runModel}
\alias{runModel}
\title{Run mosquito population model.}
\usage{
runModel(
  burnin.dates = seq(as.Date("2019-07-01"), as.Date("2020-06-30"), 1),
  burnin.reps = 100,
  run.dates = seq(as.Date("2020-07-01"), max(temp_seq$Date), 1),
  temp_seq = temp_seq[c("Date", "Tmean")],
  b = 100,
  alpha = 0.073,
  beta = 0.0315,
  K_L = 2710200,
  M_max = 1000060,
  MTD = 7.783,
  L_1 = 0,
  L_2 = 0,
  L_3 = 0,
  L_4 = 0,
  L_5 = 0,
  M = 100,
  Mfloor = 100
)
}
\arguments{
\item{burnin.dates}{Trim temperature time series to these dates for burn-in.}

\item{burnin.reps}{Repeat burn-in temperature time series n times.}

\item{run.dates}{Run temperature time series at these dates after burn-in.}

\item{temp_seq}{temperature time series}

\item{b}{number of female eggs per clutch}

\item{alpha}{adult mortality rate (1/days)}

\item{beta}{larval mortality rate (1/days)}

\item{K_L}{larval carrying capacity (numbers/km^2)}

\item{M_max}{max adult density}

\item{MTD}{minimum temperature for mosquito development)}

\item{L_1}{Initial instar densities (1-5)}

\item{L_2}{see above}

\item{L_3}{see above}

\item{L_4}{see above}

\item{L_5}{see above}

\item{M}{Initial adult density}

\item{Mfloor}{Minimum adult density}
}
\value{

}
\description{
Run mosquito population model.
}
\examples{
library(mosqmod)

# load stored temperatures
temperatures <- data.frame(mosqmod::saved_station_temps)
temperatures <- subset(temperatures, Station == "Dunedin, Musselburgh Ews")

# append temperatures (requires 'cliflo_usrid' & 'cliflo_pwd' environment
# variables to be set if not using request = FALSE)
temp_seq <- mosqmod::append_TempSeq(temp_stored = temperatures, request = FALSE)

# format sequence (fill gaps, format dates etc. )
temp_seq <- formatTempSeq(d = temp_seq)

# get projected temperatures from calendar day mean temperatures

calendar_day_mean_temps <- getCalendarDayMeans(temp_seq)

temp_projected <- project_TempSeq(temp_seq = temp_seq, extend_days = 365*2,
                                  lookback_days = 90,
                                  calendar_day_mean_temps)

# add projected temperatures
temp_seq <- dplyr::bind_rows(temp_seq, temp_projected)

# run model
res <- runModel(temp_seq = temp_seq,
                burnin.dates = seq(as.Date("2016-07-01"), as.Date("2017-06-30"), 1),
                run.dates = seq(as.Date("2017-07-01"), max(temp_seq$Date), 1))
\dontrun{
mosqmod::plot_popn(resdf = res, include_temp = TRUE, selectPopn = "M")
}
}
